Total refactor of Beeper
Rather than generating the full audio clip (16 pulses of 1000 samples
each) in a single go, buffer the `zero` and `one` pulses and emit them
in order.

Also add a dummy "header" pulse, to work around a problem I've had with
my circuit and/or Arduino not reading the first pulse correctly. This
goes along with a corresponding change to the arduino code (incoming in
a future commit) to detect and ignore the longer header pulse.