Tweak durations; add header pulse check
The "header" pulse is an extra-long pulse that begins transmission of
the remaining 16 pulses (generated by commit bb718dbb to the Android
SoundTest app); this header is detected and ignored (and sometimes, at
least on my hardware, it doesn't go through at all, which is OK because
it would be ignored anyway).

This ends up making my 16-bit pulse transmission from Android to Arduino
more reliable. YMMV.